#5 more routing fixes/changes
