+++ src/components/portal/Category/Mugic.js
@@ -100,13 +100,17 @@ export default class Mugic extends React.Component {
       );
     });
 
+    const tribes = ["Danian", "Mipedian", "OverWorld", "UnderWorld"].map((tribe, i) => (
+      <Route key={i} path={`${this.props.match.url}/${tribe}/:card`} component={SingleMugic} />
+    ));
+
     return (<div className="entry mugic">
       <div className="left">
         <div className="title">{path[2]}<hr /></div>
         {output}
       </div>
       <div className="right">
-        <Route path={`${this.props.match.url}/:card`} component={SingleMugic} />
+        {tribes}
       </div>
     </div>);
   }
